Extracted `MockAsyncSequence`
